@@ -3,12 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "react-scripts": "0.7.0"
+    "react-scripts": "0.9.5"
   },
   "dependencies": {
     "found": "../..",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2"
+    "prop-types": "^15.5.6",
+    "react": "^15.5.3",
+    "react-dom": "^15.5.3"
   },
   "scripts": {
     "start": "react-scripts start"
