@@ -6,19 +6,20 @@
     "start": "babel-node src/server"
   },
   "dependencies": {
-    "babel-cli": "^6.23.0",
-    "babel-loader": "^6.3.0",
-    "babel-preset-latest": "^6.22.0",
-    "babel-preset-react": "^6.23.0",
-    "express": "^4.14.1",
+    "babel-cli": "^6.24.1",
+    "babel-loader": "^6.4.1",
+    "babel-preset-latest": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "express": "^4.15.2",
     "farce": "^0.2.1",
     "found": "../..",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-redux": "^5.0.2",
+    "prop-types": "^15.5.6",
+    "react": "^15.5.3",
+    "react-dom": "^15.5.3",
+    "react-redux": "^5.0.4",
     "redux": "^3.6.0",
     "serialize-javascript": "^1.3.0",
-    "webpack": "^2.2.1",
-    "webpack-dev-middleware": "^1.10.0"
+    "webpack": "^2.3.3",
+    "webpack-dev-middleware": "^1.10.1"
   }
 }
