@@ -6,15 +6,16 @@
     "start": "babel-node src/server"
   },
   "dependencies": {
-    "babel-cli": "^6.18.0",
-    "babel-loader": "^6.2.8",
-    "babel-preset-latest": "^6.16.0",
-    "babel-preset-react": "^6.16.0",
-    "express": "^4.14.0",
+    "babel-cli": "^6.24.1",
+    "babel-loader": "^6.4.1",
+    "babel-preset-latest": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "express": "^4.15.2",
     "found": "../..",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
-    "webpack": "^2.1.0-beta.27",
-    "webpack-dev-middleware": "^1.8.4"
+    "prop-types": "^15.5.6",
+    "react": "^15.5.3",
+    "react-dom": "^15.5.3",
+    "webpack": "^2.3.3",
+    "webpack-dev-middleware": "^1.10.1"
   }
 }
