@@ -31,14 +31,15 @@
   },
   "homepage": "https://github.com/4Catalyzer/found#readme",
   "dependencies": {
-    "babel-runtime": "^6.18.0",
+    "babel-runtime": "^6.23.0",
     "farce": "^0.2.1",
-    "invariant": "^2.2.1",
+    "invariant": "^2.2.2",
     "is-promise": "^2.1.0",
-    "lodash": "^4.16.6",
-    "path-to-regexp": "^1.6.0",
+    "lodash": "^4.17.4",
+    "path-to-regexp": "^1.7.0",
+    "prop-types": "^15.5.6",
     "react-prop-types": "^0.4.0",
-    "react-redux": "^5.0.0-beta.3",
+    "react-redux": "^5.0.4",
     "react-static-container": "^1.0.1",
     "redux": "^3.6.0",
     "warning": "^3.0.0"
@@ -47,25 +48,25 @@
     "react": ">=0.14.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.18.0",
-    "babel-eslint": "^7.1.0",
-    "babel-jest": "^17.0.2",
+    "babel-cli": "^6.24.1",
+    "babel-eslint": "^7.2.1",
+    "babel-jest": "^19.0.0",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-dev-expression": "^0.2.1",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-latest": "^6.16.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-1": "^6.16.0",
-    "enzyme": "^2.6.0",
-    "eslint": "^3.9.1",
-    "eslint-config-4catalyzer-react": "^0.1.3",
-    "eslint-plugin-import": "^1.16.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-latest": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-1": "^6.24.1",
+    "enzyme": "^2.8.0",
+    "eslint": "^3.19.0",
+    "eslint-config-4catalyzer-react": "^0.2.0",
+    "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^2.2.3",
-    "eslint-plugin-react": "^6.5.0",
-    "jest": "^17.0.3",
-    "react": "^15.4.1",
-    "react-addons-test-utils": "^15.4.1",
-    "react-dom": "^15.4.1",
-    "rimraf": "^2.5.4"
+    "eslint-plugin-react": "^6.10.3",
+    "jest": "^19.0.2",
+    "react": "^15.5.3",
+    "react-addons-test-utils": "^15.5.1",
+    "react-dom": "^15.5.3",
+    "rimraf": "^2.6.1"
   }
 }
