@@ -1,6 +1,7 @@
 import createBrowserRouter from 'found/lib/createBrowserRouter';
 import Link from 'found/lib/Link';
 import Redirect from 'found/lib/Redirect';
+import PropTypes from 'prop-types';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
@@ -17,7 +18,7 @@ function LinkItem(props) {
 }
 
 const propTypes = {
-  children: React.PropTypes.node,
+  children: PropTypes.node,
 };
 
 function App({ children }) {
