@@ -1,34 +1,34 @@
-import React from 'react';
+import PropTypes from 'prop-types';
 
 // The shape might be different with a custom matcher or history enhancer, but
 // the default matcher assumes and provides this shape. As such, this validator
 // is purely for user convenience and should not be used internally.
-export const matchShape = React.PropTypes.shape({
-  location: React.PropTypes.shape({
-    pathname: React.PropTypes.string.isRequired,
-    query: React.PropTypes.object.isRequired,
+export const matchShape = PropTypes.shape({
+  location: PropTypes.shape({
+    pathname: PropTypes.string.isRequired,
+    query: PropTypes.object.isRequired,
   }).isRequired,
-  params: React.PropTypes.object.isRequired,
+  params: PropTypes.object.isRequired,
 });
 
 // User code generally shouldn't need this, but it doesn't hurt to export here,
 // since we use it for routerShape below.
-export const matcherShape = React.PropTypes.shape({
-  match: React.PropTypes.func.isRequired,
-  getRoutes: React.PropTypes.func.isRequired,
-  isActive: React.PropTypes.func.isRequired,
-  format: React.PropTypes.func.isRequired,
+export const matcherShape = PropTypes.shape({
+  match: PropTypes.func.isRequired,
+  getRoutes: PropTypes.func.isRequired,
+  isActive: PropTypes.func.isRequired,
+  format: PropTypes.func.isRequired,
 });
 
-export const routerShape = React.PropTypes.shape({
-  push: React.PropTypes.func.isRequired,
-  replace: React.PropTypes.func.isRequired,
-  go: React.PropTypes.func.isRequired,
+export const routerShape = PropTypes.shape({
+  push: PropTypes.func.isRequired,
+  replace: PropTypes.func.isRequired,
+  go: PropTypes.func.isRequired,
 
-  createHref: React.PropTypes.func.isRequired,
-  createLocation: React.PropTypes.func.isRequired,
-  isActive: React.PropTypes.func.isRequired,
+  createHref: PropTypes.func.isRequired,
+  createLocation: PropTypes.func.isRequired,
+  isActive: PropTypes.func.isRequired,
   matcher: matcherShape.isRequired,
 
-  addTransitionHook: React.PropTypes.func.isRequired,
+  addTransitionHook: PropTypes.func.isRequired,
 });
